fixed their code here
