+++ client/oidc-clean-client/.env
@@ -1,0 +1,15 @@
+REACT_APP_CLIENT_ID="crXYIYg2cJiNTaw5t-peoPzCRo-3JATNfBd5A86U8t0"
+REACT_APP_REDIRECT_URI="http://localhost:3000/callback"
+
+CLIENT_ASSERTION_TYPE="urn:ietf:params:oauth:client-assertion-type:jwt-bearer"
+PRIVATE_KEY_BASE64="base64_encoded_jwk"
+
+CLIENT_ID="crXYIYg2cJiNTaw5t-peoPzCRo-3JATNfBd5A86U8t0"
+REDIRECT_URI="http://localhost:3000/callback"
+REACT_APP_AUTHORIZATION_ENDPOINT="https://esignet.ida.fayda.et/authorize"
+TOKEN_ENDPOINT="https://esignet.ida.fayda.et/v1/esignet/oauth/v2/token"
+USERINFO_ENDPOINT="https://esignet.ida.fayda.et/v1/esignet/oidc/userinfo"
+
+
+EXPIRATION_TIME="15"
+ALGORITHM="RS256"
