+++ client/oidc-clean-client/public/logo512.png
