+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "client": "npm start --prefix client",
-    "server": "npm start --prefix server",
+    "client": "cd client/oidc-clean-client && npm run start",
+    "server": "npm run dev --prefix server",
     "dev": "concurrently \"npm run server\" \"npm run client\""
   },
   "keywords": [],
