+++ server/package.json
@@ -15,6 +15,6 @@
     "jose": "^5.3.0"
   },
   "devDependencies": {
-    "nodemon": "^3.0.2"
+    "nodemon": "^3.1.10"
   }
 }
